# contracts/openapi/renx-api.yaml
openapi: 3.0.3
info:
  title: RenX AI Trading Platform API
  description: |
    Comprehensive API specification for RenX AI Trading Platform
    
    ## Features
    - Multi-tenant architecture with tenant isolation
    - Role-based access control (RBAC)
    - Real-time trading and market data
    - AI-powered trading signals and predictions
    - WebSocket support for live updates
    
    ## Authentication
    All authenticated endpoints require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Multi-Tenancy
    All tenant-specific endpoints include tenant context in headers or request body.
    
  version: 1.0.0
  contact:
    name: RenX Development Team
    url: https://github.com/renx/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.renx.trading/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /trades:
    get:
      tags:
        - Trading
      summary: Get trade history
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Trade history
          content:
            application/json:
              schema:
                type: object
                properties:
                  trades:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trade'
                  total:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Trading
      summary: Place new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradeOrder'
      responses:
        '201':
          description: Trade placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
        '400':
          $ref: '#/components/responses/BadRequest'

  # Watchlist Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        orderType:
          type: string
          enum: [market, limit]
        status:
          type: string
          enum: [pending, filled, cancelled, rejected]
        createdAt:
          type: string
          format: date-time
        filledAt:
          type: string
          format: date-time
    TradeOrder:
      type: object
      required:
        - symbol
        - side
        - quantity
        - orderType
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        price:
          type: number
        orderType:
          type: string
          enum: [market, limit]
        timeInForce:
          type: string
          enum: [day, gtc, ioc, fok]
          default: day

    # Watchlist Schemas
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        message:
          type: string
        error:
          type: string
        code:
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...